move out of runner folder and add documentation
